web/zenon/setup/webapp/pyproject.toml
[tool.poetry]
name = "zenon"
version = "0.1.0"
description = ""
authors = ["Antreas Pogiatzis"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
jinja2 = "^3.1.2"
fastapi-utils = "^0.2.1"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
secweb = "^1.7.2"
requests = "^2.31.0"
cffi = "^1.15.1"
pytest-playwright = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
